+++ internal/server/api/apimodels/olsrd.go
@@ -11,7 +11,7 @@ type OlsrdLinks struct {
 type OlsrdLinkinfo struct {
 	HelloTime           uint64  `json:"helloTime"`
 	LostLinkTime        uint64  `json:"lostLinkTime"`
-	LinkQuality         uint64  `json:"linkQuality"`
+	LinkQuality         float32 `json:"linkQuality"`
 	VTime               uint64  `json:"vtime"`
 	LinkCost            float32 `json:"linkCost"`
 	LinkType            string  `json:"linkType"`
